Assign to intermediate variable rather than to sub's argument.
Within a subroutine, assigning to a hash passed to that sub as an argument is
a recipe for confusion.  In compile(), create an intermediate variable for
clearer handling of the object file.  Use a ternary rather than an if-else,
for terser code.  Restore use of arg_object_file() now that we have debugged.