use can_run from EUMM rather than IPC::Cmd
IPC::Cmd brings in a number of extra dependencies that are irrelevant to
the can_run function, which is the only part we are using. Internally,
IPC::Cmd::can_run calls functions in ExtUtils::MakeMaker. Newer versions
of EUMM provide a can_run method that provides the same functionality as
IPC::Cmd's. Just use the function from EUMM instead of IPC::Cmd.